requirements.txt
pytz==2018.9  # https://github.com/stub42/pytz

# Django
# ------------------------------------------------------------------------------
django==2.2.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-filter==2.1.0
django-allauth==0.39.1  # https://github.com/pennersr/django-allauth

django-cors-headers==3.0.2  # https://github.com/ottoyiu/django-cors-headers

# Django REST Framework
djangorestframework==3.9.1  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client
pyyaml==5.1  # https://github.com/yaml/pyyaml
django-oauth-toolkit==1.3.3
oauthlib==3.1.0

# utils
requests==2.21.0
python-dateutil==2.8.0 # https://github.com/dateutil/dateutil
ipython==7.25.0

# PostgreSQL
psycopg2-binary==2.9.1

# Testing
# ------------------------------------------------------------------------------
pytest==5.2.0  # https://github.com/pytest-dev/pytest
pytest-sugar==0.9.2  # https://github.com/Frozenball/pytest-sugar
pytest-cov==2.6.1  # https://github.com/pytest-dev/pytest-cov
httpretty==0.9.6  # https://github.com/gabrielfalcao/HTTPretty
pretend==1.0.9  # https://github.com/alex/pretend
mixer==6.1.3 # https://github.com/klen/mixero
pytest-env==0.6.2 # https://github.com/MobileDynasty/pytest-env